app/public/requirements.txt
google-cloud-bigquery
streamlit
db-dtypes
matplotlib